+++ src/main/java/com/grocery/model/Wishlist.java
@@ -19,7 +19,7 @@ public class Wishlist {
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE)
 	@Column(name = "wishlist_id")
-	private int wishlistId;
+	private Long wishlistId;
 	
 	@OneToOne
 	private User user;
